Destruct main resizin source to more separated sources
