+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 node_js:
 - 8
-script: npm run lint && npm test && npm i && npm run build
+script: npm run check && npm test && npm i && npm run build
 deploy:
   provider: npm
   email: ackeedevelopment@gmail.com
