+++ tsconfig.json
@@ -10,8 +10,11 @@
 		// Don't emit; allow Babel to transform files.
 		"noEmit": true,
 		// Enable strictest settings like strictNullChecks & noImplicitAny.
-		"strict": true,
-		// Disallow features that require cross-file information for emit.
+
+        // TODO - disabled strict until end of refactor phase
+        "strict": false,
+
+        // Disallow features that require cross-file information for emit.
 		"isolatedModules": true,
 		// Import non-ES modules as default imports.
         "esModuleInterop": true,
